content is being saved and added
my content from my form is being added and saved, last time I pushed I managed to put the content in an array en put it in a json file, but it would overwrite the content it had. So I googled on how to add data to a json file and found it.