retrieve title data
I can retrieve the title data, but when I do it in the view.php it doesn't work at all, if I do it in the index it works, but this is not what we want. Can not figure out what I'm doing wrong. Big challenge this one, not liking PHP at the moment. But we'll get there.